Migrate majority of code to Userspace
