+++ users/drashna/rules.mk
@@ -1,0 +1,2 @@
+
+SRC += drashna.c
